// src/app/types/CharacterType.ts
export type CharacterType = {
  debut: {
    appearsIn: string;
  };
  images: string[];
  jutsu: string[];
  name: string;
  personal: {
    status: string;
    birthdate: string;
    clan: string;
    kekkeiGenkai: string[];
    classification: string[];
  };
  natureType: string[];
  tools: string[];
};